requirements.txt
aiohttp>=3.11.18
python-dotenv>=1.1.0
python-telegram-bot>=20.10
